// libs/enums/faq.enum.ts
export enum FaqStatus {
  ACTIVE = 'ACTIVE',
  BLOCKED = 'BLOCKED',
  DELETED = 'DELETED'
  }

  export enum FaqCategory {
    GENERAL = 'GENERAL',
    FOR_BUYERS = 'FOR_BUYERS',
    FOR_STORES = 'FOR_STORES',
    MEMBERSHIP = 'MEMBERSHIP',
    COMMUNITY = 'COMMUNITY',
    PAYMENT = 'PAYMENT',
    PRODUCT = 'PRODUCT',
    OTHERS = 'OTHERS', 
  }

  